+++ sentry-android-core/src/main/java/io/sentry/android/core/ManifestMetadataReader.java
@@ -71,6 +71,7 @@ final class ManifestMetadataReader {
 
   static final String ATTACH_SCREENSHOT = "io.sentry.attach-screenshot";
   static final String CLIENT_REPORTS_ENABLE = "io.sentry.send-client-reports";
+  static final String COLLECT_ADDITIONAL_CONTEXT = "io.sentry.additional-context";
 
   /** ManifestMetadataReader ctor */
   private ManifestMetadataReader() {}
@@ -209,6 +210,13 @@ final class ManifestMetadataReader {
 
         options.setSendClientReports(
             readBool(metadata, logger, CLIENT_REPORTS_ENABLE, options.isSendClientReports()));
+
+        options.setCollectAdditionalContext(
+            readBool(
+                metadata,
+                logger,
+                COLLECT_ADDITIONAL_CONTEXT,
+                options.isCollectAdditionalContext()));
 
         if (options.getTracesSampleRate() == null) {
           final Double tracesSampleRate = readDouble(metadata, logger, TRACES_SAMPLE_RATE);