+++ CHANGELOG.md
@@ -2,9 +2,41 @@
 
 ## Unreleased
 
+### Fixes
+
+- Only send userid in Dynamic Sampling Context if sendDefaultPii is true ([#2147](https://github.com/getsentry/sentry-java/pull/2147))
+
+### Features
+
+- New package `sentry-android-navigation` for AndroidX Navigation support ([#2136](https://github.com/getsentry/sentry-java/pull/2136))
+- New package `sentry-compose` for Jetpack Compose support (Navigation) ([#2136](https://github.com/getsentry/sentry-java/pull/2136))
+- Add sample rate to baggage as well as trace in envelope header and flatten user ([#2135](https://github.com/getsentry/sentry-java/pull/2135))
+
+## 6.1.4
+
+### Fixes
+
+- Filter out app starts with more than 60s ([#2127](https://github.com/getsentry/sentry-java/pull/2127))
+
+## 6.1.3
+
+### Fixes
+
+- Fix thread leak due to Timer being created and never cancelled ([#2131](https://github.com/getsentry/sentry-java/pull/2131))
+
+## 6.1.2
+
+### Fixes
+
+- Swallow error when reading ActivityManager#getProcessesInErrorState instead of crashing ([#2114](https://github.com/getsentry/sentry-java/pull/2114))
+- Use charset string directly as StandardCharsets is not available on earlier Android versions ([#2111](https://github.com/getsentry/sentry-java/pull/2111))
+
+## 6.1.1
+
 ### Features
 
 - Replace `tracestate` header with `baggage` header ([#2078](https://github.com/getsentry/sentry-java/pull/2078))
+- Allow opting out of device info collection that requires Inter-Process Communication (IPC) ([#2100](https://github.com/getsentry/sentry-java/pull/2100))
 
 ## 6.1.0
 
@@ -53,6 +85,7 @@
             - Call `toString()` where object graphs exceed max depth.
     - Remove `gson` dependency.
     - Remove `IUnknownPropertiesConsumer`
+- Pass MDC tags as Sentry tags ([#1954](https://github.com/getsentry/sentry-java/pull/1954))
 
 ### Fixes